http://code.google.com/p/netsentry/issues/detail?id=11 Use long instead of int to parse the byte counters. This raises the limit from about 1.9gb to 8192pb.

+++ trunk/src/com/googlecode/netsentry/backend/Updater.java
@@ -422,8 +422,8 @@ public class Updater extends BroadcastReceiver {
 
             if (matcher.matches()) {
                 String deviceName = matcher.group(1).trim();
-                int bytesReceived = Integer.parseInt(matcher.group(2));
-                int bytesSent = Integer.parseInt(matcher.group(10));
+                long bytesReceived = Long.parseLong(matcher.group(2));
+                long bytesSent = Long.parseLong(matcher.group(10));
                 InterfaceStats stats = Updater.sInterfacesStatsMap.get(deviceName);
 
                 if (stats == null) {
